+++ template-base-3-typescript/template/src/components/common/DefaultView.tsx
@@ -1,4 +1,4 @@
-import Header from 'components/common/Header';
+import Header from './header/Header';
 
 export default function DefaultView({ children }: { children: JSX.Element }) {
   return (
